azure-card-analyzer/requirements.txt
streamlit
python-dotenv
azure-core
azure-storage-blob
azure-ai-documentintelligence